requirements.txt
transformers==4.46.0
hydra-core
pytest
debugpy
datasets
accelerate
loguru
sentencepiece
einops
pandas
tqdm
huggingface_hub